Remove comma from enabled_backends
As there is no way to correctly pick up existing
enabled_backends from Fuel we have no need to add
the backends parameters with a comma to the
enabled_backends that is being set.

If you have 2 Cinder plugins enabled then the last one
to run overwrite enabled_backends anyway. This needs
to be investigated by Mirantis

Change-Id: Ifeb72c2bad5285ef11d524c078dc62cebbb758f8