Get checks passing
First changes: Make the puppet-lint tool happy

We skip a handful of checks, but at least now should catch some of the more simple stuff. This also formats the files a little more consistently.

Second change: Update the docs makefile so we can actually build the pdf with the expected commands

Change-Id: Ia8e9224cad38baf5e4602f9be7af828dcffe6bbb